requirements.txt
requests>=2.25.1
click>=7.1.2
python-dotenv>=0.15.0
spacy==3.7.5
scispacy==0.5.4
https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_core_sci_md-0.5.4.tar.gz
colorlog>=6.7.0
pytrials==0.3.0
openai>=1.0.0
pandas>=1.5.0
nicegui>=1.4.0
flask>=2.0.0
uuid>=1.30